refactor(web): rename src→web, fetch manifest at boot
Frontend now boots from fetch('/api/manifest') instead of reading the
__MANIFEST__ placeholder out of an embedded <script> tag. Drops the
readEmbeddedJson helper, the EMBEDDED_MANIFEST DOM id, and the
__PROJECT_NAME__ title placeholder along with it.

JS tests move to web/tests/ (colocated). main.test.js is gone — its only
job was exercising readEmbeddedJson. Bash test scripts deleted; Python
test_build.py deleted (build.py itself was already removed in the move).

package.json simplified: drop vite-plugin-singlefile, drop test:scripts /
test:drift; rename to "codecity"; type:module to match the ESM source.
