--- src/index.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>CodeCity — __PROJECT_NAME__</title>
-</head>
-<body>
-  <div id="tree-sidebar"></div>
-  <canvas id="city"></canvas>
-  <div id="sidebar"></div>
-
-  <!-- Three.js is fetched from CDN at open-time (not bundled) so the output
-       HTML stays small. Cached by the browser across codecity runs. -->
-  <script type="importmap">
-  {
-    "imports": {
-      "three":         "https://cdn.jsdelivr.net/npm/three@0.161.0/build/three.module.min.js",
-      "three/addons/": "https://cdn.jsdelivr.net/npm/three@0.161.0/examples/jsm/"
-    }
-  }
-  </script>
-
-  <!-- MANIFEST is injected as JSON text here.
-         dev:  vite plugin fills the token with the dev-fixture manifest.
-         build: vite leaves the token in place; codecity.py replaces it at
-                skill runtime. The visual config now lives in src/defaults.js
-                and is bundled into the JS module. -->
-  <script type="application/json" id="codecity-manifest">__MANIFEST__</script>
-
-  <script type="module" src="./main.js"></script>
-</body>
-</html>
