+++ vitest.config.js
@@ -2,7 +2,7 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    include: ['tests/**/*.test.js'],
+    include: ['web/tests/**/*.test.js'],
     environment: 'jsdom',
   },
 });
