+++ web/config/view.js
