+++ web/index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>CodeCity</title>
+</head>
+<body>
+  <div id="tree-sidebar"></div>
+  <canvas id="city"></canvas>
+  <div id="sidebar"></div>
+
+  <!-- Three.js is fetched from CDN at open-time (not bundled). Cached by the
+       browser across codecity runs. -->
+  <script type="importmap">
+  {
+    "imports": {
+      "three":         "https://cdn.jsdelivr.net/npm/three@0.161.0/build/three.module.min.js",
+      "three/addons/": "https://cdn.jsdelivr.net/npm/three@0.161.0/examples/jsm/"
+    }
+  }
+  </script>
+
+  <script type="module" src="./main.js"></script>
+</body>
+</html>
