+++ web/tests/scene/hsl.test.js
@@ -5,7 +5,7 @@ import {
   shadeColor,
   shadeAndShiftHue,
   shadeByRatio,
-} from '../../src/scene/hsl.js';
+} from '../../scene/hsl.js';
 
 describe('hslToComponents', () => {
   it('parses a standard hsl() string', () => {
